renamed folders to represent actual file names
